Make tables of responsibilities work
So far the right-most column did not work correctly. When  an
owner is selected in the column to the left of it, the admins
and watchers of that selected owner should be displayed.

And the first row should be auto-selected when the
responsibilities-tab is selected.

Both issues are fixed by this commit.