Set focus to user input field on login form
In order to be able to immediately start typing (or pasting
credentials from a password manager), the input field for
the user name should have focus immediately.

This is achieved here and also implemented in the callback
that is called when the input mask is switched from
email/password to LDAP-login (using ref and $refs).

This fixes #48.