Switch Tabulator from version 4.9.3 to 5.6.1
Major version upgrade necessary for functionality and
performance reasons.
Mainly for the general service overview which takes minutes
to load with 4.9.3 and produces Tabulator errors in the console
and bugs in the UI.
All that is fixed in 5.6.1.

Major change: from synchronous initialization to asynchronous
initialization.
This makes listening to tableBuilt event mandatory before
filling a table with data (e.g. with setData-function).